+++ packages/grafana-ui/src/components/BigValue/BigValue.tsx
@@ -16,7 +16,7 @@ export interface BigValueSparkline {
   maxX: number;
 }
 
-export enum SingleStatDisplayMode {
+export enum BigValueDisplayMode {
   Classic,
   Classic2,
   Vibrant,
@@ -30,7 +30,7 @@ export interface Props extends Themeable {
   sparkline?: BigValueSparkline;
   onClick?: React.MouseEventHandler<HTMLElement>;
   className?: string;
-  displayMode: SingleStatDisplayMode;
+  displayMode: BigValueDisplayMode;
 }
 
 export class BigValue extends PureComponent<Props> {
@@ -73,7 +73,7 @@ interface LayoutResult {
   type: LayoutType;
   width: number;
   height: number;
-  displayMode: SingleStatDisplayMode;
+  displayMode: BigValueDisplayMode;
   theme: GrafanaTheme;
   valueColor: string;
 }
@@ -135,9 +135,9 @@ export function calculateLayout(props: Props): LayoutResult {
   }
 
   switch (displayMode) {
-    case SingleStatDisplayMode.Vibrant2:
-    case SingleStatDisplayMode.Classic:
-    case SingleStatDisplayMode.Classic2:
+    case BigValueDisplayMode.Vibrant2:
+    case BigValueDisplayMode.Classic:
+    case BigValueDisplayMode.Classic2:
       chartWidth = width;
       chartHeight += PANEL_PADDING;
       break;
@@ -181,8 +181,8 @@ export function getValueStyles(layout: LayoutResult) {
   };
 
   switch (layout.displayMode) {
-    case SingleStatDisplayMode.Classic:
-    case SingleStatDisplayMode.Classic2:
+    case BigValueDisplayMode.Classic:
+    case BigValueDisplayMode.Classic2:
       styles.color = layout.valueColor;
   }
 
@@ -234,8 +234,8 @@ export function getPanelStyles(layout: LayoutResult) {
   const themeFactor = layout.theme.isDark ? 1 : -0.7;
 
   switch (layout.displayMode) {
-    case SingleStatDisplayMode.Vibrant:
-    case SingleStatDisplayMode.Vibrant2:
+    case BigValueDisplayMode.Vibrant:
+    case BigValueDisplayMode.Vibrant2:
       const bgColor2 = tinycolor(layout.valueColor)
         .darken(15 * themeFactor)
         .spin(8)
@@ -246,8 +246,8 @@ export function getPanelStyles(layout: LayoutResult) {
         .toRgbString();
       panelStyles.background = `linear-gradient(120deg, ${bgColor2}, ${bgColor3})`;
       break;
-    case SingleStatDisplayMode.Classic:
-    case SingleStatDisplayMode.Classic2:
+    case BigValueDisplayMode.Classic:
+    case BigValueDisplayMode.Classic2:
       panelStyles.background = `${layout.theme.colors.dark4}`;
       break;
   }
@@ -318,13 +318,13 @@ function renderGraph(layout: LayoutResult, sparkline?: BigValueSparkline) {
   }
 
   switch (layout.displayMode) {
-    case SingleStatDisplayMode.Vibrant2:
+    case BigValueDisplayMode.Vibrant2:
       geomRender = renderVibrant2Geom;
       break;
-    case SingleStatDisplayMode.Classic:
+    case BigValueDisplayMode.Classic:
       geomRender = renderClassicAreaGeom;
       break;
-    case SingleStatDisplayMode.Classic2:
+    case BigValueDisplayMode.Classic2:
       geomRender = renderAreaGeom;
       break;
   }