+++ public/app/plugins/panel/stat/module.tsx
@@ -1,12 +1,12 @@
 import { sharedSingleStatMigrationHandler, sharedSingleStatPanelChangedHandler } from '@grafana/ui';
 import { PanelPlugin } from '@grafana/data';
-import { SingleStatOptions, defaults } from './types';
-import { SingleStatPanel } from './SingleStatPanel';
-import { SingleStatEditor } from './SingleStatEditor';
+import { StatPanelOptions, defaults } from './types';
+import { StatPanel } from './StatPanel';
+import { StatPanelEditor } from './StatPanelEditor';
 
-export const plugin = new PanelPlugin<SingleStatOptions>(SingleStatPanel)
+export const plugin = new PanelPlugin<StatPanelOptions>(StatPanel)
   .setDefaults(defaults)
-  .setEditor(SingleStatEditor)
+  .setEditor(StatPanelEditor)
   .setNoPadding()
   .setPanelChangeHandler(sharedSingleStatPanelChangedHandler)
   .setMigrationHandler(sharedSingleStatMigrationHandler);